requirements.txt
anthropic>=0.49.0
rumps>=0.4.0
pyobjc-framework-Quartz>=10.0
